+++ .github/workflows/docker-wiki-update.yml
@@ -37,12 +37,12 @@ jobs:
         uses: actions/checkout@v4
         with:
           repository: ${{ github.repository }}.wiki
-          path: wiki/
+          path: wiki_src/
 
       - name: Update wiki 🏷
         run: >
-          python3 -m tagging.update_wiki
-          --wiki-dir wiki/
+          python3 -m wiki.update_wiki
+          --wiki-dir wiki_src/
           --hist-lines-dir /tmp/jupyter/hist_lines/
           --manifests-dir /tmp/jupyter/manifests/
           --repository ${{ github.repository }}
