+++ docs/maintaining/tagging_examples/git_helper.py
@@ -1,0 +1,4 @@
+from tagging.utils.git_helper import GitHelper
+
+print("Git hash:", GitHelper.commit_hash())
+print("Git message:", GitHelper.commit_message())
