+++ tagging/apps/write_manifest.py
@@ -7,12 +7,15 @@ from pathlib import Path
 
 from docker.models.containers import Container
 
-from tagging.common_arguments import common_arguments_parser
-from tagging.docker_runner import DockerRunner
-from tagging.get_prefix import get_file_prefix, get_tag_prefix
-from tagging.get_taggers_and_manifests import get_taggers_and_manifests
-from tagging.git_helper import GitHelper
-from tagging.manifests import ManifestHeader, ManifestInterface
+from tagging.apps.common_cli_arguments import common_arguments_parser
+from tagging.hierarchy.get_taggers_and_manifests import (
+    get_taggers_and_manifests,
+)
+from tagging.manifests.header import ManifestHeader
+from tagging.manifests.manifest_interface import ManifestInterface
+from tagging.utils.docker_runner import DockerRunner
+from tagging.utils.get_prefix import get_file_prefix, get_tag_prefix
+from tagging.utils.git_helper import GitHelper
 
 LOGGER = logging.getLogger(__name__)
