+++ tagging/hierarchy/get_taggers_and_manifests.py
@@ -1,9 +1,9 @@
 # Copyright (c) Jupyter Development Team.
 # Distributed under the terms of the Modified BSD License.
 
-from tagging.images_hierarchy import ALL_IMAGES
-from tagging.manifests import ManifestInterface
-from tagging.taggers import TaggerInterface
+from tagging.hierarchy.images_hierarchy import ALL_IMAGES
+from tagging.manifests.manifest_interface import ManifestInterface
+from tagging.taggers.tagger_interface import TaggerInterface
 
 
 def get_taggers_and_manifests(
