+++ tagging/taggers/tagger_interface.py
@@ -1,0 +1,9 @@
+from docker.models.containers import Container
+
+
+class TaggerInterface:
+    """Common interface for all taggers"""
+
+    @staticmethod
+    def tag_value(container: Container) -> str:
+        raise NotImplementedError
