+++ tagging/taggers/versions.py
@@ -1,11 +1,9 @@
 # Copyright (c) Jupyter Development Team.
 # Distributed under the terms of the Modified BSD License.
-import datetime
-
 from docker.models.containers import Container
 
-from tagging.docker_runner import DockerRunner
-from tagging.git_helper import GitHelper
+from tagging.taggers.tagger_interface import TaggerInterface
+from tagging.utils.docker_runner import DockerRunner
 
 
 def _get_program_version(container: Container, program: str) -> str:
@@ -23,39 +21,6 @@ def _get_pip_package_version(container: Container, package: str) -> str:
     version_line = package_info.split("\n")[1]
     assert version_line.startswith(PIP_VERSION_PREFIX)
     return version_line[len(PIP_VERSION_PREFIX) :]
-
-
-class TaggerInterface:
-    """Common interface for all taggers"""
-
-    @staticmethod
-    def tag_value(container: Container) -> str:
-        raise NotImplementedError
-
-
-class SHATagger(TaggerInterface):
-    @staticmethod
-    def tag_value(container: Container) -> str:
-        return GitHelper.commit_hash_tag()
-
-
-class DateTagger(TaggerInterface):
-    @staticmethod
-    def tag_value(container: Container) -> str:
-        return datetime.datetime.now(datetime.UTC).strftime("%Y-%m-%d")
-
-
-class UbuntuVersionTagger(TaggerInterface):
-    @staticmethod
-    def tag_value(container: Container) -> str:
-        os_release = DockerRunner.run_simple_command(
-            container,
-            "cat /etc/os-release",
-        ).split("\n")
-        for line in os_release:
-            if line.startswith("VERSION_ID"):
-                return "ubuntu-" + line.split("=")[1].strip('"')
-        raise RuntimeError(f"did not find ubuntu version in: {os_release}")
 
 
 class PythonVersionTagger(TaggerInterface):
